KroneckerArraysTensorAlgebraExt extension
